auth con Google implementado. aut para horarios, cambiar coordina de profesor y genericDeleteBttn implementados

+++ src/components/GenericDeleteBttn.tsx
@@ -2,6 +2,7 @@
 
 import React from "react";
 import { useRouter } from 'next/navigation';
+import { useSessionContext, useUser } from '@supabase/auth-helpers-react';
 
 interface DeleteButtonProps {
     id: string;
@@ -9,10 +10,15 @@ interface DeleteButtonProps {
 }
 
 export default function GenericDeleteBttn({id, api}: DeleteButtonProps){
+    const { isLoading } = useSessionContext();
+    const user = useUser();
     const router = useRouter();
+
+    if (isLoading) return <div>Cargando...</div>;
 
     return(
         <button className="btn btn-danger" type="button" onClick={async () => {
+          if (!user) return <div>No autenticado</div>;
             try {
                 const response = await fetch('/api/'+api, {
                   method: 'DELETE',
